+++ index.html
@@ -38,13 +38,10 @@
 
   <button class="archiveButton">Show archive rates</button>
 
-  <ul class = "archive-rates-list">
-
-  </div>
-  <div 
+  <ul class = "archive-rates-list"></ul>
 </div>
 
-<script type="module" src="script.js"></script>
+<script type="module" src="src/app.js"></script>
 
 </body>
 </html>
