+++ src/constants.js
@@ -1,0 +1,3 @@
+const ADDRESS = 'https://api.frankfurter.app/latest';
+
+export { ADDRESS };
